+++ apps/admin/src/views/SettingsPage.vue
@@ -8,11 +8,11 @@
         <ion-card-title>Seguridad</ion-card-title>
       </ion-card-header>
       <ion-card-content>
-        <ion-item lines="none" :disabled="supportsBiometrics">
+        <ion-item lines="none" :disabled="!supportsBiometrics">
           <ion-label>Activar incio de sesión biométrico</ion-label>
           <ion-toggle
-            :disabled="supportsBiometrics"
-            v-model="enableBiometrics"
+            :disabled="!supportsBiometrics"
+            v-model="biometricsAreEnabled"
             @ionChange="toggleBiometrics($event)"
           ></ion-toggle>
         </ion-item>
@@ -36,7 +36,7 @@
           :disabled="!isPlatform('capacitor')"
         >
           <ion-label>Cambiar permisos</ion-label>
-          <ion-icon slot="end" :md="open" :ios="openOutline"></ion-icon>
+          <ion-icon slot="end" :icon="openOutline"></ion-icon>
         </ion-item>
         <router-link to="/dev/logs">
           <ion-item lines="none" button :detail="true">
@@ -57,7 +57,6 @@
 </template>
 
 <script lang="ts" setup>
-import { user } from '../lib/user';
 import {
   IonTitle,
   IonCard,
@@ -69,7 +68,6 @@ import {
   IonToggle,
   IonButton,
   IonIcon,
-  toastController,
   alertController,
   modalController,
   useIonRouter,
@@ -77,69 +75,118 @@ import {
 } from '@ionic/vue';
 import { NativeBiometric } from 'capacitor-native-biometric';
 import DevSettingsModal from '../components/DevSettingsModal.vue';
-import { open, openOutline } from 'ionicons/icons';
+import { openOutline, checkmarkCircleOutline } from 'ionicons/icons';
 import {
   NativeSettings,
   AndroidSettings,
   IOSSettings,
 } from 'capacitor-native-settings';
-import PageWrapper from '../lib/components/PageWrapper.vue';
+import PageWrapper from '../components/PageWrapper.vue';
+import { getUser } from '../user';
 
 const router = useIonRouter();
+const logger = useLogger();
 
-const supportsBiometrics = ref(true);
-const enableBiometrics = ref(false);
+const user = getUser();
+
+const supportsBiometrics = ref(false);
+const biometricsAreEnabled = ref(false);
 
 if (isPlatform('capacitor'))
   NativeBiometric.isAvailable().then(
-    ({ isAvailable }) => (supportsBiometrics.value = !isAvailable)
+    ({ isAvailable }) => (supportsBiometrics.value = isAvailable)
   );
 
 // TODO in login you have to also check if biometrics is available, imagine that someone toggles on the option and the disables biometrics, it should't prompt them for biometrics
 Preferences.get({
   key: 'ceebiAdmin.biometricsEnabled',
 }).then((val) => {
-  if (val.value || '' !== '') enableBiometrics.value = true;
+  if (val.value) biometricsAreEnabled.value = true;
 });
 
 const toggleBiometrics = async (ev: any) => {
   console.info(ev);
   // In case they are toggling on the option
-  if (enableBiometrics.value) {
-    const verified = await NativeBiometric.verifyIdentity({
-      reason: 'Verifica tu identidad para activar esta opción',
-      title: 'Verifica tu identidad para activar esta opción',
-    })
-      .then(() => true)
-      .catch(() => false);
+  if (supportsBiometrics.value) {
+    if (ev.target.checked) {
+      const verified = await NativeBiometric.verifyIdentity({
+        reason: 'Verifica tu identidad para activar esta opción',
+        title: 'Verifica tu identidad para activar esta opción',
+      })
+        .then(() => true)
+        .catch(() => false);
 
-    // If the user couldn't be verified, notify them and revert setting
-    if (!verified) {
-      alertController
-        .create({
-          message: 'No ha sido posible verificar la identidad',
-          buttons: ['Aceptar'],
-        })
-        .then((alert) => alert.present());
-      enableBiometrics.value = false;
-    } else {
-      // User has been succesfully verified, so save login info for next time and save preference
-      NativeBiometric.setCredentials({
-        username: user.value?.username || '',
-        password: user.value?.password || '',
-        server: 'es.biociencias.admin',
-      }).then(() =>
-        toastController
+      // If the user couldn't be verified, notify them and revert setting
+      if (!verified) {
+        alertController
           .create({
-            message: 'Ajustes de autenticación guardados',
-            duration: 1500,
+            message: 'No ha sido posible verificar la identidad',
+            buttons: ['Aceptar'],
           })
-          .then((toast) => toast.present())
-      );
-      Preferences.set({
+          .then((alert) => alert.present());
+        biometricsAreEnabled.value = false;
+      } else {
+        const prompt = await alertController.create({
+          header: 'Contraseña',
+          message:
+            'Introduce tu contraseña para guardarla en el archivo seguro de claves',
+          inputs: [
+            {
+              type: 'password',
+              placeholder: 'Contraseña',
+            },
+          ],
+          buttons: [
+            {
+              text: 'Cancelar',
+              role: 'cancel',
+            },
+            {
+              text: 'Guardar',
+              handler: ({ 0: input_password }: { 0: string }) => {
+                if (input_password === '') return;
+                // User has been succesfully verified, so save login info for next time and save preference
+                NativeBiometric.setCredentials({
+                  username: user.value?.wp.email || '',
+                  password: input_password,
+                  server: 'es.biociencias.admin',
+                }).then(() =>
+                  useToast({
+                    message: 'Ajustes de autenticación guardados',
+                    color: 'success',
+                    icon: checkmarkCircleOutline,
+                  })
+                );
+                Preferences.set({
+                  key: 'ceebiAdmin.biometricsEnabled',
+                  value: 'true',
+                });
+              },
+            },
+          ],
+        });
+        prompt.present();
+        console.log('eh?');
+      }
+    } else {
+      // TODO Delete data from android's biometric keystore
+      Preferences.remove({
         key: 'ceebiAdmin.biometricsEnabled',
-        value: 'true',
-      });
+      })
+        .then(() =>
+          useToast({
+            message: 'Inicio de sesión biométrico desactivado',
+            color: 'success',
+            icon: checkmarkCircleOutline,
+          })
+        )
+        .catch((error) => {
+          logger.error(
+            'settings:toggleBiometrics',
+            'error when deactivating biometrics',
+            { error }
+          );
+        });
     }
   }
 };
@@ -167,6 +214,14 @@ const openPermissions = () => {
     optionIOS: IOSSettings.App,
   });
 };
+
+watch(supportsBiometrics, (val) => {
+  logger.trace(
+    'settings:watch(supportsBiometrics)',
+    'changed supports biometrics',
+    val
+  );
+});
 </script>
 
 <style scoped>