+++ apps/admin/src/views/UserDetailsPage.vue
@@ -84,14 +84,18 @@
           >{{ isAdmin ? 'Admin' : 'User' }}</IonBadge
         >
       </ion-card-header>
-      <IonCardContent>
+      <IonCardContent v-if="isAdmin">
         <!-- TODO Add help button to show up an alert explaining each permission -->
+        <!-- TODO Add an onclick.prevent.stop for allow_schedule and show toast explaining that it is yet to implement -->
         <IonItem>
           <IonLabel> Editar franjas de asistencia </IonLabel>
           <IonToggle
             slot="end"
             type="email"
-            :disabled="user?.supabase.allow_admins"
+            @ion-change="
+              updatePermissions('allow_schedule', $event.target.checked)
+            "
+            :disabled="user?.supabase.allow_admins || true"
             :checked="user?.supabase.allow_schedule || false"
           ></IonToggle>
         </IonItem>
@@ -100,7 +104,10 @@
           <IonToggle
             slot="end"
             type="email"
-            :disabled="user?.supabase.allow_admins || false"
+            @ion-change="
+              updatePermissions('allow_attendance', $event.target.checked)
+            "
+            :disabled="user?.supabase.allow_admins || true"
             :checked="user?.supabase.allow_attendance || false"
           ></IonToggle>
         </IonItem>
@@ -109,16 +116,22 @@
           <IonToggle
             slot="end"
             type="email"
-            :disabled="user?.supabase.allow_admins"
+            @ion-change="
+              updatePermissions('allow_notifications', $event.target.checked)
+            "
+            :disabled="user?.supabase.allow_admins || true"
             :checked="user?.supabase.allow_notifications || false"
           ></IonToggle>
         </IonItem>
         <IonItem>
-          <IonLabel> Ver y editar usuarios </IonLabel>
+          <IonLabel> Ver los datos de todos los usuarios </IonLabel>
           <IonToggle
             slot="end"
             type="email"
-            :disabled="user?.supabase.allow_admins"
+            @ion-change="
+              updatePermissions('allow_users', $event.target.checked)
+            "
+            :disabled="user?.supabase.allow_admins || true"
             :checked="user?.supabase.allow_users || false"
           ></IonToggle>
         </IonItem>
@@ -127,9 +140,27 @@
           <IonToggle
             slot="end"
             type="email"
-            :disabled="user?.supabase.allow_admins"
+            @ion-change="
+              updatePermissions('allow_admins', $event.target.checked)
+            "
+            :disabled="user?.supabase.allow_admins || true"
             :checked="user?.supabase.allow_admins || false"
           ></IonToggle>
+        </IonItem>
+      </IonCardContent>
+      <IonCardContent v-else>
+        <IonItem
+          @click="updatePermissions('is_admin', true)"
+          button
+          lines="none"
+          class="text-cblue"
+        >
+          Convertir usuario en administrador
+          <IonIcon
+            class="text-inherit"
+            :icon="arrowUpCircleOutline"
+            slot="end"
+          ></IonIcon>
         </IonItem>
       </IonCardContent>
     </ion-card>
@@ -166,9 +197,17 @@ import {
   IonLabel,
   IonBadge,
   IonToggle,
+  IonIcon,
 } from '@ionic/vue';
-import { alertCircleOutline, code } from 'ionicons/icons';
+import {
+  alertCircleOutline,
+  arrowUpCircleOutline,
+  checkmarkCircleOutline,
+  closeCircleOutline,
+  code,
+} from 'ionicons/icons';
 
+import { FirebaseCrashlytics } from '@capacitor-community/firebase-crashlytics';
 import { useCacheable } from '../cache';
 import type { DefinedCacheable } from '../cache';
 import { Database } from 'libs/supabase/src/types';
@@ -176,6 +215,8 @@ import { wpapi } from '../req';
 import { authHeaders } from '../wpauth';
 
 const route = useRoute();
+const supabase = useSupabase();
+const logger = useLogger();
 
 const apiResponse = ref();
 
@@ -207,7 +248,7 @@ const { value: user, isLoading } = (
     apiResponse.value = res;
     const wpUser = await res.json<WPUser>();
 
-    const { data: supabaseUser, error } = await useSupabase()
+    const { data: supabaseUser, error } = await supabase
       .from('users')
       .select('*')
       .eq('id', wpUser.acf.id_base_de_datos_app)
@@ -237,9 +278,6 @@ const { value: user, isLoading } = (
   {
     email: '',
     id: userId.value as string,
-    first_name: (userName.value as string).split(' ', 1)[0],
-    last_name: (userName.value as string).split(' ', 1)[1],
-    id_document: '',
     is_admin: false,
     allow_notifications: false,
     allow_users: false,
@@ -247,7 +285,7 @@ const { value: user, isLoading } = (
     allow_assistance: false, // TODO This is attendance not assistance
     allow_register: false,
     allow_admins: false,
-  } as Database['public']['Tables']['profiles']['Row']
+  } as Database['public']['Tables']['users']['Row']
 );
 
 const isAdmin = computed(() =>
@@ -263,6 +301,46 @@ watch(userId, () => {
 watch(isLoading, (val) => {
   if (!val) loadingController.dismiss();
 });
+
+const updatePermissions = (
+  key:
+    | 'is_admin'
+    | 'allow_admins'
+    | 'allow_attendance'
+    | 'allow_notifications'
+    | 'allow_schedule'
+    | 'allow_users',
+  status: boolean
+) => {
+  supabase
+    .from('users')
+    .update({
+      [key]: status,
+    })
+    .eq('id', user.value?.supabase.id)
+    .then(({ data, error }) => {
+      if (error) {
+        useToast({
+          message: 'Error actualizando usuario, error reportado',
+          color: 'danger',
+          icon: closeCircleOutline,
+        });
+        logger.error(
+          'userDetails:updatePermissions',
+          'error from supabase on update',
+          { data, error }
+        );
+        // FirebaseCrashlytics.recordException({
+        //   stacktrace: StackTrace,
+        // });
+      } else
+        useToast({
+          message: 'Usuario actualizado',
+          color: 'success',
+          icon: checkmarkCircleOutline,
+        });
+    });
+};
 
 const openJSON = () => {
   modalController